Updated the clone tests
I think that we need to move the arrays to a mock. I don't know what it needs.

Also I put the jar that is required in the artifacts folder so that we can all just add it.

My last test doesn't work, its the one which clones an empty 2D array.